refactor: uses Buttons comp and replaces on Terminal page
