+++ packages/evmcrispr-terminal/src/pages/terminal/useTerminalStore.ts
